requirements.txt
flask==3.1.0
scikit-learn==1.6.1
pandas==2.2.3
requests==2.32.3
joblib==1.4.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
